Fixing Ramonas fuck Up's

+++ src/main/kotlin/codeacademy/bookingforum/app/comment/CommentDto.kt
@@ -6,8 +6,8 @@ data class CommentDto(
     val id : Long? = null,
     val dateCommented : LocalDateTime? = null,
     val content : String? = "",
-    val postId : Int? = null,
-    val userId : Int? = null,
+    val postId : Long? = null,
+    val userId : Long? = null,
     val likes : Int? = 0,
     val dislikes : Int? = 0,
 )
